zmiany kosmetyczne

+++ zad3/fileread.py
@@ -5,6 +5,6 @@ def read(path):
     with open(path) as f:
         array = f.readline().split(',')
         for i in array:
-            x.append(int(i))
+            x.append(float(i))
 
         return x
